ggv2-deploy-cloud/artifacts/requirements.txt
wheel
Cython
numpy==1.19.5
opencv-python-headless==4.5.3.56
dlr==1.8.0
awsiotsdk